:muscle: Split is.ts into multiple files under `is` directory
