+++ is/core_test.ts
@@ -1,0 +1,223 @@
+import {
+  assertEquals,
+  assertStrictEquals,
+} from "https://deno.land/std@0.211.0/assert/mod.ts";
+import { stringify } from "./_testutil.ts";
+import { type Predicate } from "./type.ts";
+import is, {
+  isAny,
+  isArray,
+  isAsyncFunction,
+  isBigInt,
+  isBoolean,
+  isFunction,
+  isMap,
+  isNull,
+  isNullish,
+  isNumber,
+  isPrimitive,
+  isRecord,
+  isSet,
+  isString,
+  isSymbol,
+  isSyncFunction,
+  isUndefined,
+  isUnknown,
+} from "./core.ts";
+
+const examples = {
+  string: ["", "Hello world"],
+  number: [0, 1234567890],
+  bigint: [0n, 1234567890n],
+  boolean: [true, false],
+  array: [[], [0, 1, 2], ["a", "b", "c"], [0, "a", true]],
+  set: [new Set(), new Set([0, 1, 2]), new Set(["a", "b", "c"])],
+  record: [{}, { a: 0, b: 1, c: 2 }, { a: "a", b: "b", c: "c" }],
+  map: [
+    new Map(),
+    new Map([["a", 0], ["b", 1], ["c", 2]]),
+    new Map([["a", "a"], ["b", "b"], ["c", "c"]]),
+  ],
+  syncFunction: [function a() {}, () => {}],
+  asyncFunction: [async function b() {}, async () => {}],
+  null: [null],
+  undefined: [undefined],
+  symbol: [Symbol("a"), Symbol("b"), Symbol("c")],
+  date: [new Date(1690248225000), new Date(0)],
+  promise: [new Promise(() => {})],
+} as const;
+
+async function testWithExamples<T>(
+  t: Deno.TestContext,
+  pred: Predicate<T>,
+  opts?: {
+    validExamples?: (keyof typeof examples)[];
+    excludeExamples?: (keyof typeof examples)[];
+  },
+): Promise<void> {
+  const { validExamples = [], excludeExamples = [] } = opts ?? {};
+  const exampleEntries = (Object.entries(examples) as unknown as [
+    name: keyof typeof examples,
+    example: unknown[],
+  ][]).filter(([k]) => !excludeExamples.includes(k));
+  for (const [name, example] of exampleEntries) {
+    const expect = validExamples.includes(name);
+    for (const v of example) {
+      await t.step(
+        `returns ${expect} on ${stringify(v)}`,
+        () => {
+          assertEquals(pred(v), expect);
+        },
+      );
+    }
+  }
+}
+
+Deno.test("isAny", async (t) => {
+  await testWithExamples(t, isAny, {
+    validExamples: [
+      "string",
+      "number",
+      "bigint",
+      "boolean",
+      "array",
+      "set",
+      "record",
+      "map",
+      "syncFunction",
+      "asyncFunction",
+      "null",
+      "undefined",
+      "symbol",
+      "date",
+      "promise",
+    ],
+  });
+});
+
+Deno.test("isUnknown", async (t) => {
+  await testWithExamples(t, isUnknown, {
+    validExamples: [
+      "string",
+      "number",
+      "bigint",
+      "boolean",
+      "array",
+      "set",
+      "record",
+      "map",
+      "syncFunction",
+      "asyncFunction",
+      "null",
+      "undefined",
+      "symbol",
+      "date",
+      "promise",
+    ],
+  });
+});
+
+Deno.test("isString", async (t) => {
+  await testWithExamples(t, isString, { validExamples: ["string"] });
+});
+
+Deno.test("isNumber", async (t) => {
+  await testWithExamples(t, isNumber, { validExamples: ["number"] });
+});
+
+Deno.test("isBigInt", async (t) => {
+  await testWithExamples(t, isBigInt, { validExamples: ["bigint"] });
+});
+
+Deno.test("isBoolean", async (t) => {
+  await testWithExamples(t, isBoolean, { validExamples: ["boolean"] });
+});
+
+Deno.test("isArray", async (t) => {
+  await testWithExamples(t, isArray, { validExamples: ["array"] });
+});
+
+Deno.test("isSet", async (t) => {
+  await testWithExamples(t, isSet, { validExamples: ["set"] });
+});
+
+Deno.test("isRecord", async (t) => {
+  await testWithExamples(t, isRecord, {
+    validExamples: ["record"],
+  });
+});
+
+Deno.test("isMap", async (t) => {
+  await testWithExamples(t, isMap, {
+    validExamples: ["map"],
+  });
+});
+
+Deno.test("isFunction", async (t) => {
+  await testWithExamples(t, isFunction, {
+    validExamples: ["syncFunction", "asyncFunction"],
+  });
+});
+
+Deno.test("isSyncFunction", async (t) => {
+  await testWithExamples(t, isSyncFunction, {
+    validExamples: ["syncFunction"],
+  });
+});
+
+Deno.test("isAsyncFunction", async (t) => {
+  await testWithExamples(t, isAsyncFunction, {
+    validExamples: ["asyncFunction"],
+  });
+});
+
+Deno.test("isNull", async (t) => {
+  await testWithExamples(t, isNull, { validExamples: ["null"] });
+});
+
+Deno.test("isUndefined", async (t) => {
+  await testWithExamples(t, isUndefined, { validExamples: ["undefined"] });
+});
+
+Deno.test("isNullish", async (t) => {
+  await testWithExamples(t, isNullish, {
+    validExamples: ["null", "undefined"],
+  });
+});
+
+Deno.test("isSymbol", async (t) => {
+  await testWithExamples(t, isSymbol, { validExamples: ["symbol"] });
+});
+
+Deno.test("isPrimitive", async (t) => {
+  await testWithExamples(t, isPrimitive, {
+    validExamples: [
+      "string",
+      "number",
+      "bigint",
+      "boolean",
+      "null",
+      "undefined",
+      "symbol",
+    ],
+  });
+});
+
+Deno.test("is", async (t) => {
+  const mod = await import("./core.ts");
+  const casesOfAliasAndIsFunction = Object.entries(mod)
+    .filter(([k, _]) => k.startsWith("is"))
+    .map(([k, v]) => [k.slice(2), v] as const);
+  for (const [alias, fn] of casesOfAliasAndIsFunction) {
+    await t.step(`defines \`${alias}\` function`, () => {
+      assertStrictEquals(is[alias as keyof typeof is], fn);
+    });
+  }
+  await t.step(
+    "only has entries that are the same as the `is*` function aliases",
+    () => {
+      const aliases = casesOfAliasAndIsFunction.map(([a]) => a).sort();
+      assertEquals(Object.keys(is).sort(), aliases);
+    },
+  );
+});
