+++ is/factory_test.ts
@@ -6,48 +6,23 @@ import {
   assertSnapshot,
 } from "https://deno.land/std@0.211.0/testing/snapshot.ts";
 import { assertType } from "https://deno.land/std@0.211.0/testing/types.ts";
+import { type Equal, stringify } from "./_testutil.ts";
+import { type Predicate } from "./type.ts";
+import { isArray, isBoolean, isNumber, isString } from "./core.ts";
 import is, {
-  isAllOf,
-  isAny,
-  isArray,
   isArrayOf,
-  isAsyncFunction,
-  isBigInt,
-  isBoolean,
-  isFunction,
   isInstanceOf,
   isLiteralOf,
   isLiteralOneOf,
-  isMap,
   isMapOf,
-  isNull,
-  isNullish,
-  isNumber,
   isObjectOf,
-  isOneOf,
-  isOptionalOf,
-  isPrimitive,
   isReadonlyTupleOf,
   isReadonlyUniformTupleOf,
-  isRecord,
   isRecordOf,
-  isSet,
   isSetOf,
-  isString,
-  isSymbol,
-  isSyncFunction,
   isTupleOf,
-  isUndefined,
   isUniformTupleOf,
-  isUnknown,
-  Predicate,
-  PredicateType,
-} from "./is.ts";
-
-// It seems 'IsExact' in deno_std is false positive so use `Equal` in type-challenges
-// https://github.com/type-challenges/type-challenges/blob/e77262dba62e9254451f661cb4fe5517ffd1d933/utils/index.d.ts#L7-L9
-type Equal<X, Y> = (<T>() => T extends X ? 1 : 2) extends
-  (<T>() => T extends Y ? 1 : 2) ? true : false;
+} from "./factory.ts";
 
 const examples = {
   string: ["", "Hello world"],
@@ -70,15 +45,6 @@ const examples = {
   date: [new Date(1690248225000), new Date(0)],
   promise: [new Promise(() => {})],
 } as const;
-
-function stringify(x: unknown): string {
-  if (x instanceof Date) return `Date(${x.valueOf()})`;
-  if (x instanceof Promise) return "Promise";
-  if (typeof x === "function") return x.toString();
-  if (typeof x === "bigint") return `${x}n`;
-  if (typeof x === "symbol") return x.toString();
-  return JSON.stringify(x);
-}
 
 async function testWithExamples<T>(
   t: Deno.TestContext,
@@ -106,91 +72,6 @@ async function testWithExamples<T>(
   }
 }
 
-Deno.test("PredicateType", () => {
-  const isArticle = is.ObjectOf({
-    title: is.String,
-    body: is.String,
-    refs: is.ArrayOf(is.OneOf([
-      is.String,
-      is.ObjectOf({
-        name: is.String,
-        url: is.String,
-      }),
-    ])),
-  });
-  assertType<
-    Equal<PredicateType<typeof isArticle>, {
-      title: string;
-      body: string;
-      refs: (string | { name: string; url: string })[];
-    }>
-  >(true);
-});
-
-Deno.test("isAny", async (t) => {
-  await testWithExamples(t, isAny, {
-    validExamples: [
-      "string",
-      "number",
-      "bigint",
-      "boolean",
-      "array",
-      "set",
-      "record",
-      "map",
-      "syncFunction",
-      "asyncFunction",
-      "null",
-      "undefined",
-      "symbol",
-      "date",
-      "promise",
-    ],
-  });
-});
-
-Deno.test("isUnknown", async (t) => {
-  await testWithExamples(t, isUnknown, {
-    validExamples: [
-      "string",
-      "number",
-      "bigint",
-      "boolean",
-      "array",
-      "set",
-      "record",
-      "map",
-      "syncFunction",
-      "asyncFunction",
-      "null",
-      "undefined",
-      "symbol",
-      "date",
-      "promise",
-    ],
-  });
-});
-
-Deno.test("isString", async (t) => {
-  await testWithExamples(t, isString, { validExamples: ["string"] });
-});
-
-Deno.test("isNumber", async (t) => {
-  await testWithExamples(t, isNumber, { validExamples: ["number"] });
-});
-
-Deno.test("isBigInt", async (t) => {
-  await testWithExamples(t, isBigInt, { validExamples: ["bigint"] });
-});
-
-Deno.test("isBoolean", async (t) => {
-  await testWithExamples(t, isBoolean, { validExamples: ["boolean"] });
-});
-
-Deno.test("isArray", async (t) => {
-  await testWithExamples(t, isArray, { validExamples: ["array"] });
-});
-
 Deno.test("isArrayOf<T>", async (t) => {
   await t.step("returns properly named function", async (t) => {
     await assertSnapshot(t, isArrayOf(isNumber).name);
@@ -215,10 +96,6 @@ Deno.test("isArrayOf<T>", async (t) => {
   await testWithExamples(t, isArrayOf((_: unknown): _ is unknown => true), {
     excludeExamples: ["array"],
   });
-});
-
-Deno.test("isSet", async (t) => {
-  await testWithExamples(t, isSet, { validExamples: ["set"] });
 });
 
 Deno.test("isSetOf<T>", async (t) => {
@@ -289,11 +166,11 @@ Deno.test("isTupleOf<T, E>", async (t) => {
   await t.step("returns properly named function", async (t) => {
     await assertSnapshot(
       t,
-      isTupleOf([isNumber, isString, isBoolean], is.Array).name,
+      isTupleOf([isNumber, isString, isBoolean], isArray).name,
     );
     await assertSnapshot(
       t,
-      isTupleOf([(_x): _x is string => false], is.ArrayOf(is.String))
+      isTupleOf([(_x): _x is string => false], isArrayOf(isString))
         .name,
     );
     // Nested
@@ -301,15 +178,15 @@ Deno.test("isTupleOf<T, E>", async (t) => {
       t,
       isTupleOf([
         isTupleOf(
-          [isTupleOf([isNumber, isString, isBoolean], is.Array)],
-          is.Array,
+          [isTupleOf([isNumber, isString, isBoolean], isArray)],
+          isArray,
         ),
       ]).name,
     );
   });
   await t.step("returns proper type predicate", () => {
     const predTup = [isNumber, isString, isBoolean] as const;
-    const predElse = is.ArrayOf(is.Number);
+    const predElse = isArrayOf(isNumber);
     const a: unknown = [0, "a", true, 0, 1, 2];
     if (isTupleOf(predTup, predElse)(a)) {
       assertType<Equal<typeof a, [number, string, boolean, ...number[]]>>(
@@ -319,12 +196,12 @@ Deno.test("isTupleOf<T, E>", async (t) => {
   });
   await t.step("returns true on T tuple", () => {
     const predTup = [isNumber, isString, isBoolean] as const;
-    const predElse = is.ArrayOf(is.Number);
+    const predElse = isArrayOf(isNumber);
     assertEquals(isTupleOf(predTup, predElse)([0, "a", true, 0, 1, 2]), true);
   });
   await t.step("returns false on non T tuple", () => {
     const predTup = [isNumber, isString, isBoolean] as const;
-    const predElse = is.ArrayOf(is.String);
+    const predElse = isArrayOf(isString);
     assertEquals(isTupleOf(predTup, predElse)([0, 1, 2, 0, 1, 2]), false);
     assertEquals(isTupleOf(predTup, predElse)([0, "a", 0, 1, 2]), false);
     assertEquals(
@@ -333,7 +210,7 @@ Deno.test("isTupleOf<T, E>", async (t) => {
     );
     assertEquals(isTupleOf(predTup, predElse)([0, "a", true, 0, 1, 2]), false);
   });
-  const predElse = is.Array;
+  const predElse = isArray;
   await testWithExamples(
     t,
     isTupleOf([(_: unknown): _ is unknown => true], predElse),
@@ -391,14 +268,14 @@ Deno.test("isReadonlyTupleOf<T, E>", async (t) => {
   await t.step("returns properly named function", async (t) => {
     await assertSnapshot(
       t,
-      isReadonlyTupleOf([isNumber, isString, isBoolean], is.Array)
+      isReadonlyTupleOf([isNumber, isString, isBoolean], isArray)
         .name,
     );
     await assertSnapshot(
       t,
       isReadonlyTupleOf(
         [(_x): _x is string => false],
-        is.ArrayOf(is.String),
+        isArrayOf(isString),
       ).name,
     );
     // Nested
@@ -406,14 +283,14 @@ Deno.test("isReadonlyTupleOf<T, E>", async (t) => {
       t,
       isReadonlyTupleOf([
         isReadonlyTupleOf([
-          isReadonlyTupleOf([isNumber, isString, isBoolean], is.Array),
-        ], is.Array),
-      ], is.Array).name,
+          isReadonlyTupleOf([isNumber, isString, isBoolean], isArray),
+        ], isArray),
+      ], isArray).name,
     );
   });
   await t.step("returns proper type predicate", () => {
     const predTup = [isNumber, isString, isBoolean] as const;
-    const predElse = is.ArrayOf(is.Number);
+    const predElse = isArrayOf(isNumber);
     const a: unknown = [0, "a", true, 0, 1, 2];
     if (isReadonlyTupleOf(predTup, predElse)(a)) {
       assertType<
@@ -423,7 +300,7 @@ Deno.test("isReadonlyTupleOf<T, E>", async (t) => {
   });
   await t.step("returns true on T tuple", () => {
     const predTup = [isNumber, isString, isBoolean] as const;
-    const predElse = is.ArrayOf(is.Number);
+    const predElse = isArrayOf(isNumber);
     assertEquals(
       isReadonlyTupleOf(predTup, predElse)([0, "a", true, 0, 1, 2]),
       true,
@@ -431,7 +308,7 @@ Deno.test("isReadonlyTupleOf<T, E>", async (t) => {
   });
   await t.step("returns false on non T tuple", () => {
     const predTup = [isNumber, isString, isBoolean] as const;
-    const predElse = is.ArrayOf(is.String);
+    const predElse = isArrayOf(isString);
     assertEquals(
       isReadonlyTupleOf(predTup, predElse)([0, 1, 2, 0, 1, 2]),
       false,
@@ -449,7 +326,7 @@ Deno.test("isReadonlyTupleOf<T, E>", async (t) => {
       false,
     );
   });
-  const predElse = is.Array;
+  const predElse = isArray;
   await testWithExamples(
     t,
     isReadonlyTupleOf([(_: unknown): _ is unknown => true], predElse),
@@ -484,12 +361,12 @@ Deno.test("isUniformTupleOf<T>", async (t) => {
   });
   await t.step("returns true on mono-typed T tuple", () => {
     assertEquals(isUniformTupleOf(3)([0, 1, 2]), true);
-    assertEquals(isUniformTupleOf(3, is.Number)([0, 1, 2]), true);
+    assertEquals(isUniformTupleOf(3, isNumber)([0, 1, 2]), true);
   });
   await t.step("returns false on non mono-typed T tuple", () => {
     assertEquals(isUniformTupleOf(4)([0, 1, 2]), false);
     assertEquals(isUniformTupleOf(4)([0, 1, 2, 3, 4]), false);
-    assertEquals(isUniformTupleOf(3, is.Number)(["a", "b", "c"]), false);
+    assertEquals(isUniformTupleOf(3, isNumber)(["a", "b", "c"]), false);
   });
   await testWithExamples(t, isUniformTupleOf(4), {
     excludeExamples: ["array"],
@@ -524,24 +401,18 @@ Deno.test("isReadonlyUniformTupleOf<T>", async (t) => {
   });
   await t.step("returns true on mono-typed T tuple", () => {
     assertEquals(isReadonlyUniformTupleOf(3)([0, 1, 2]), true);
-    assertEquals(isReadonlyUniformTupleOf(3, is.Number)([0, 1, 2]), true);
+    assertEquals(isReadonlyUniformTupleOf(3, isNumber)([0, 1, 2]), true);
   });
   await t.step("returns false on non mono-typed T tuple", () => {
     assertEquals(isReadonlyUniformTupleOf(4)([0, 1, 2]), false);
     assertEquals(isReadonlyUniformTupleOf(4)([0, 1, 2, 3, 4]), false);
     assertEquals(
-      isReadonlyUniformTupleOf(3, is.Number)(["a", "b", "c"]),
+      isReadonlyUniformTupleOf(3, isNumber)(["a", "b", "c"]),
       false,
     );
   });
   await testWithExamples(t, isReadonlyUniformTupleOf(4), {
     excludeExamples: ["array"],
-  });
-});
-
-Deno.test("isRecord", async (t) => {
-  await testWithExamples(t, isRecord, {
-    validExamples: ["record"],
   });
 });
 
@@ -602,12 +473,6 @@ Deno.test("isRecordOf<T, K>", async (t) => {
   });
   await testWithExamples(t, isRecordOf((_: unknown): _ is unknown => true), {
     excludeExamples: ["record", "date", "promise"],
-  });
-});
-
-Deno.test("isMap", async (t) => {
-  await testWithExamples(t, isMap, {
-    validExamples: ["map"],
   });
 });
 
@@ -749,118 +614,6 @@ Deno.test("isObjectOf<T>", async (t) => {
     isObjectOf({ a: (_: unknown): _ is unknown => false }),
     { excludeExamples: ["record"] },
   );
-  await t.step("with optional properties", async (t) => {
-    await t.step("returns proper type predicate", () => {
-      const predObj = {
-        a: isNumber,
-        b: isOneOf([isString, isUndefined]),
-        c: isOptionalOf(isBoolean),
-      };
-      const a: unknown = { a: 0, b: "a" };
-      if (isObjectOf(predObj)(a)) {
-        assertType<
-          Equal<typeof a, { a: number; b: string | undefined; c?: boolean }>
-        >(true);
-      }
-    });
-    await t.step("returns true on T object", () => {
-      const predObj = {
-        a: isNumber,
-        b: isOneOf([isString, isUndefined]),
-        c: isOptionalOf(isBoolean),
-      };
-      assertEquals(isObjectOf(predObj)({ a: 0, b: "a", c: true }), true);
-      assertEquals(
-        isObjectOf(predObj)({ a: 0, b: "a" }),
-        true,
-        "Object does not have an optional property",
-      );
-      assertEquals(
-        isObjectOf(predObj)({ a: 0, b: "a", c: undefined }),
-        true,
-        "Object has `undefined` as value of optional property",
-      );
-      assertEquals(
-        isObjectOf(predObj, { strict: true })({ a: 0, b: "a", c: true }),
-        true,
-        "Specify `{ strict: true }`",
-      );
-      assertEquals(
-        isObjectOf(predObj, { strict: true })({ a: 0, b: "a" }),
-        true,
-        "Specify `{ strict: true }` and object does not have one optional property",
-      );
-    });
-    await t.step("returns false on non T object", () => {
-      const predObj = {
-        a: isNumber,
-        b: isOneOf([isString, isUndefined]),
-        c: isOptionalOf(isBoolean),
-      };
-      assertEquals(
-        isObjectOf(predObj)({ a: 0, b: "a", c: "" }),
-        false,
-        "Object have a different type property",
-      );
-      assertEquals(
-        isObjectOf(predObj)({ a: 0, b: "a", c: null }),
-        false,
-        "Object has `null` as value of optional property",
-      );
-      assertEquals(
-        isObjectOf(predObj, { strict: true })({
-          a: 0,
-          b: "a",
-          c: true,
-          d: "invalid",
-        }),
-        false,
-        "Specify `{ strict: true }` and object have an unknown property",
-      );
-      assertEquals(
-        isObjectOf(predObj, { strict: true })({
-          a: 0,
-          b: "a",
-          d: "invalid",
-        }),
-        false,
-        "Specify `{ strict: true }` and object have the same number of properties but an unknown property exists",
-      );
-    });
-  });
-});
-
-Deno.test("isFunction", async (t) => {
-  await testWithExamples(t, isFunction, {
-    validExamples: ["syncFunction", "asyncFunction"],
-  });
-  assertType<
-    Equal<PredicateType<typeof isFunction>, (...args: unknown[]) => unknown>
-  >(true);
-});
-
-Deno.test("isSyncFunction", async (t) => {
-  await testWithExamples(t, isSyncFunction, {
-    validExamples: ["syncFunction"],
-  });
-  assertType<
-    Equal<
-      PredicateType<typeof isSyncFunction>,
-      (...args: unknown[]) => unknown
-    >
-  >(true);
-});
-
-Deno.test("isAsyncFunction", async (t) => {
-  await testWithExamples(t, isAsyncFunction, {
-    validExamples: ["asyncFunction"],
-  });
-  assertType<
-    Equal<
-      PredicateType<typeof isAsyncFunction>,
-      (...args: unknown[]) => Promise<unknown>
-    >
-  >(true);
 });
 
 Deno.test("isInstanceOf<T>", async (t) => {
@@ -902,38 +655,6 @@ Deno.test("isInstanceOf<T>", async (t) => {
     if (isInstanceOf(Promise)(c)) {
       assertType<Equal<typeof c, Promise<unknown>>>(true);
     }
-  });
-});
-
-Deno.test("isNull", async (t) => {
-  await testWithExamples(t, isNull, { validExamples: ["null"] });
-});
-
-Deno.test("isUndefined", async (t) => {
-  await testWithExamples(t, isUndefined, { validExamples: ["undefined"] });
-});
-
-Deno.test("isNullish", async (t) => {
-  await testWithExamples(t, isNullish, {
-    validExamples: ["null", "undefined"],
-  });
-});
-
-Deno.test("isSymbol", async (t) => {
-  await testWithExamples(t, isSymbol, { validExamples: ["symbol"] });
-});
-
-Deno.test("isPrimitive", async (t) => {
-  await testWithExamples(t, isPrimitive, {
-    validExamples: [
-      "string",
-      "number",
-      "bigint",
-      "boolean",
-      "null",
-      "undefined",
-      "symbol",
-    ],
   });
 });
 
@@ -986,169 +707,8 @@ Deno.test("isLiteralOneOf<T>", async (t) => {
   });
 });
 
-Deno.test("isOneOf<T>", async (t) => {
-  await t.step("returns properly named function", async (t) => {
-    await assertSnapshot(t, isOneOf([isNumber, isString, isBoolean]).name);
-  });
-  await t.step("returns proper type predicate", () => {
-    const preds = [isNumber, isString, isBoolean] as const;
-    const a: unknown = [0, "a", true];
-    if (isOneOf(preds)(a)) {
-      assertType<Equal<typeof a, number | string | boolean>>(true);
-    }
-  });
-  await t.step("returns proper type predicate (#49)", () => {
-    const isFoo = isObjectOf({ foo: isString });
-    const isBar = isObjectOf({ foo: isString, bar: isNumber });
-    type Foo = PredicateType<typeof isFoo>;
-    type Bar = PredicateType<typeof isBar>;
-    const preds = [isFoo, isBar] as const;
-    const a: unknown = [0, "a", true];
-    if (isOneOf(preds)(a)) {
-      assertType<Equal<typeof a, Foo | Bar>>(true);
-    }
-  });
-  await t.step("returns true on one of T", () => {
-    const preds = [isNumber, isString, isBoolean] as const;
-    assertEquals(isOneOf(preds)(0), true);
-    assertEquals(isOneOf(preds)("a"), true);
-    assertEquals(isOneOf(preds)(true), true);
-  });
-  await t.step("returns false on non of T", async (t) => {
-    const preds = [isNumber, isString, isBoolean] as const;
-    await testWithExamples(t, isOneOf(preds), {
-      excludeExamples: ["number", "string", "boolean"],
-    });
-  });
-});
-
-Deno.test("isAllOf<T>", async (t) => {
-  await t.step("returns properly named function", async (t) => {
-    await assertSnapshot(
-      t,
-      isAllOf([
-        is.ObjectOf({ a: is.Number }),
-        is.ObjectOf({ b: is.String }),
-      ]).name,
-    );
-  });
-  await t.step("returns proper type predicate", () => {
-    const preds = [
-      is.ObjectOf({ a: is.Number }),
-      is.ObjectOf({ b: is.String }),
-    ] as const;
-    const a: unknown = { a: 0, b: "a" };
-    if (isAllOf(preds)(a)) {
-      assertType<Equal<typeof a, { a: number } & { b: string }>>(true);
-    }
-  });
-  await t.step("returns true on all of T", () => {
-    const preds = [
-      is.ObjectOf({ a: is.Number }),
-      is.ObjectOf({ b: is.String }),
-    ] as const;
-    assertEquals(isAllOf(preds)({ a: 0, b: "a" }), true);
-  });
-  await t.step("returns false on non of T", async (t) => {
-    const preds = [
-      is.ObjectOf({ a: is.Number }),
-      is.ObjectOf({ b: is.String }),
-    ] as const;
-    assertEquals(
-      isAllOf(preds)({ a: 0, b: 0 }),
-      false,
-      "Some properties has wrong type",
-    );
-    assertEquals(
-      isAllOf(preds)({ a: 0 }),
-      false,
-      "Some properties does not exists",
-    );
-    await testWithExamples(t, isAllOf(preds), {
-      excludeExamples: ["record"],
-    });
-  });
-});
-
-Deno.test("isOptionalOf<T>", async (t) => {
-  await t.step("returns properly named function", async (t) => {
-    await assertSnapshot(t, isOptionalOf(isNumber).name);
-  });
-  await t.step("returns proper type predicate", () => {
-    const a: unknown = undefined;
-    if (isOptionalOf(isNumber)(a)) {
-      assertType<Equal<typeof a, number | undefined>>(true);
-    }
-  });
-  await t.step("with isString", async (t) => {
-    await testWithExamples(t, isOptionalOf(isString), {
-      validExamples: ["string", "undefined"],
-    });
-  });
-  await t.step("with isNumber", async (t) => {
-    await testWithExamples(t, isOptionalOf(isNumber), {
-      validExamples: ["number", "undefined"],
-    });
-  });
-  await t.step("with isBigInt", async (t) => {
-    await testWithExamples(t, isOptionalOf(isBigInt), {
-      validExamples: ["bigint", "undefined"],
-    });
-  });
-  await t.step("with isBoolean", async (t) => {
-    await testWithExamples(t, isOptionalOf(isBoolean), {
-      validExamples: ["boolean", "undefined"],
-    });
-  });
-  await t.step("with isArray", async (t) => {
-    await testWithExamples(t, isOptionalOf(isArray), {
-      validExamples: ["array", "undefined"],
-    });
-  });
-  await t.step("with isSet", async (t) => {
-    await testWithExamples(t, isOptionalOf(isSet), {
-      validExamples: ["set", "undefined"],
-    });
-  });
-  await t.step("with isRecord", async (t) => {
-    await testWithExamples(t, isOptionalOf(isRecord), {
-      validExamples: ["record", "undefined"],
-    });
-  });
-  await t.step("with isFunction", async (t) => {
-    await testWithExamples(t, isOptionalOf(isFunction), {
-      validExamples: ["syncFunction", "asyncFunction", "undefined"],
-    });
-  });
-  await t.step("with isSyncFunction", async (t) => {
-    await testWithExamples(t, isOptionalOf(isSyncFunction), {
-      validExamples: ["syncFunction", "undefined"],
-    });
-  });
-  await t.step("with isAsyncFunction", async (t) => {
-    await testWithExamples(t, isOptionalOf(isAsyncFunction), {
-      validExamples: ["asyncFunction", "undefined"],
-    });
-  });
-  await t.step("with isNull", async (t) => {
-    await testWithExamples(t, isOptionalOf(isNull), {
-      validExamples: ["null", "undefined"],
-    });
-  });
-  await t.step("with isUndefined", async (t) => {
-    await testWithExamples(t, isOptionalOf(isUndefined), {
-      validExamples: ["undefined"],
-    });
-  });
-  await t.step("with isSymbol", async (t) => {
-    await testWithExamples(t, isOptionalOf(isSymbol), {
-      validExamples: ["symbol", "undefined"],
-    });
-  });
-});
-
 Deno.test("is", async (t) => {
-  const mod = await import("./is.ts");
+  const mod = await import("./factory.ts");
   const casesOfAliasAndIsFunction = Object.entries(mod)
     .filter(([k, _]) => k.startsWith("is"))
     .map(([k, v]) => [k.slice(2), v] as const);
